requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
decouple==0.0.7
Django==3.2.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
Pillow==8.2.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
django-heroku==0.3.1